+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-02 11:34:40
+
+### Changed Files
+- `supabase/functions/slack-policy-notification/index.ts`
+
+### Statistics
+```
+ 1 file changed, 22 insertions(+)
+```
+
 ## [Checkpoint] - 2026-01-02 10:45:05
 
 ### Changed Files